chore: Mejora al proyecto FastAPI
