+++ config/database.py
@@ -1,0 +1,6 @@
+from pymongo import MongoClient
+
+# MongoDB connection
+client = MongoClient("mongodb://localhost:27017")
+db = client["fastapi"]
+posts_collection = db["post"]
