+++ data.js
@@ -1,22 +1,22 @@
 [
   {
-    _id: {
-      $oid: "66d7f3991d5bb2043ea157c2",
+    "_id": {
+      "$oid": "66d7f3991d5bb2043ea157c2"
     },
-    title: "First Post",
-    content: "This is the content of the first post.",
-    created: {
-      $date: "2024-09-04T05:43:53.631Z",
-    },
+    "title": "First Post",
+    "content": "This is the content of the first post.",
+    "created": {
+      "$date": "2024-09-04T05:43:53.631Z"
+    }
   },
   {
-    _id: {
-      $oid: "66d7f3991d5bb2043ea157c3",
+    "_id": {
+      "$oid": "66d7f3991d5bb2043ea157c3"
     },
-    title: "Second Post",
-    content: "This is the content of the second post.",
-    created: {
-      $date: "2024-09-04T05:43:53.631Z",
-    },
-  },
-];
+    "title": "Second Post",
+    "content": "This is the content of the second post.",
+    "created": {
+      "$date": "2024-09-04T05:43:53.631Z"
+    }
+  }
+]
